Consolidate tool modules
