+++ tools/__init__.py
@@ -1,3 +1,16 @@
 """Utility helpers for the HelpDesk API."""
 
-from .operation_result import OperationResult
+from .system_utilities import OperationResult
+from .reference_data import ReferenceDataManager
+from .ticket_management import TicketManager, TicketTools
+from .analytics_reporting import AnalyticsManager
+from .user_services import UserManager
+
+__all__ = [
+    "OperationResult",
+    "ReferenceDataManager",
+    "TicketManager",
+    "TicketTools",
+    "AnalyticsManager",
+    "UserManager",
+]
