+++ tools/asset_tools.py
@@ -1,41 +1,29 @@
-"""Database helpers for retrieving and listing asset information."""
+"""Deprecated - use :mod:`tools.reference_data` instead."""
 
+from __future__ import annotations
+
+import warnings
 from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy import select
-import logging
-
 from db.models import Asset
 
-logger = logging.getLogger(__name__)
+from .reference_data import ReferenceDataManager
+
+_manager = ReferenceDataManager()
 
 
 async def get_asset(db: AsyncSession, asset_id: int) -> Asset | None:
-    """Retrieve a single asset by its primary key.
-
-    Args:
-        db: Async SQLAlchemy session used for the query.
-        asset_id: Identifier of the asset to fetch.
-
-    Returns:
-        The ``Asset`` instance if found, otherwise ``None``.
-    """
-
-    return await db.get(Asset, asset_id)
+    warnings.warn(
+        "get_asset is deprecated; use ReferenceDataManager.get_asset",
+        DeprecationWarning,
+        stacklevel=2,
+    )
+    return await _manager.get_asset(db, asset_id)
 
 
 async def list_assets(db: AsyncSession, skip: int = 0, limit: int = 10) -> list[Asset]:
-    """Return a paginated list of assets ordered by ID.
-
-    Args:
-        db: Async SQLAlchemy session used for the query.
-        skip: Number of records to skip from the beginning.
-        limit: Maximum number of records to return.
-
-    Returns:
-        A list of ``Asset`` instances.
-    """
-
-    result = await db.execute(
-        select(Asset).order_by(Asset.ID).offset(skip).limit(limit)
+    warnings.warn(
+        "list_assets is deprecated; use ReferenceDataManager.list_assets",
+        DeprecationWarning,
+        stacklevel=2,
     )
-    return list(result.scalars().all())
+    return await _manager.list_assets(db, skip=skip, limit=limit)
