+++ tools/attachment_tools.py
@@ -1,29 +1,20 @@
+"""Deprecated - use :mod:`tools.ticket_management` instead."""
 
+from __future__ import annotations
 
-"""Helpers for working with ticket attachments."""
-
-
+import warnings
 from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy import select
-import logging
-
 from db.models import TicketAttachment
 
-logger = logging.getLogger(__name__)
+from .ticket_management import TicketManager
+
+_manager = TicketManager()
 
 
 async def get_ticket_attachments(db: AsyncSession, ticket_id: int) -> list[TicketAttachment]:
-    """Return all attachments associated with a ticket.
-
-    Args:
-        db: Async SQLAlchemy session used for the query.
-        ticket_id: Identifier of the ticket whose attachments are requested.
-
-    Returns:
-        A list of ``TicketAttachment`` instances.
-    """
-
-    result = await db.execute(
-        select(TicketAttachment).filter(TicketAttachment.Ticket_ID == ticket_id)
+    warnings.warn(
+        "get_ticket_attachments is deprecated; use TicketManager.get_attachments",
+        DeprecationWarning,
+        stacklevel=2,
     )
-    return list(result.scalars().all())
+    return await _manager.get_attachments(db, ticket_id)
