+++ tools/oncall_tools.py
@@ -1,37 +1,31 @@
-"""Helpers for retrieving on-call shift information."""
+"""Deprecated - use :mod:`tools.user_services` instead."""
 
 from __future__ import annotations
 
-from datetime import datetime, UTC
-import logging
-from typing import Sequence
-
-from sqlalchemy import select
+import warnings
 from sqlalchemy.ext.asyncio import AsyncSession
+from typing import Sequence
 
 from db.models import OnCallShift
 
-logger = logging.getLogger(__name__)
+from .user_services import UserManager
+
+_manager = UserManager()
 
 
 async def get_current_oncall(db: AsyncSession) -> OnCallShift | None:
-    """Return the on-call shift active at the current time."""
-    now = datetime.now(UTC)
-    result = await db.execute(
-        select(OnCallShift)
-        .where(OnCallShift.start_time <= now)
-        .where(OnCallShift.end_time > now)
-        .order_by(OnCallShift.start_time.desc())
-        .limit(1)
+    warnings.warn(
+        "get_current_oncall is deprecated; use UserManager.get_current_oncall",
+        DeprecationWarning,
+        stacklevel=2,
     )
-    shift = result.scalars().first()
-    logger.info("Current on-call shift: %s", shift)
-    return shift
+    return await _manager.get_current_oncall(db)
 
 
 async def list_oncall_schedule(db: AsyncSession, skip: int = 0, limit: int = 10) -> Sequence[OnCallShift]:
-    """Return upcoming on-call shifts ordered by start_time."""
-    result = await db.execute(
-        select(OnCallShift).order_by(OnCallShift.start_time).offset(skip).limit(limit)
+    warnings.warn(
+        "list_oncall_schedule is deprecated; use UserManager.list_oncall_schedule",
+        DeprecationWarning,
+        stacklevel=2,
     )
-    return result.scalars().all()
+    return await _manager.list_oncall_schedule(db, skip=skip, limit=limit)
