+++ tools/site_tools.py
@@ -1,42 +1,29 @@
-"""Utilities for reading site information from the database."""
+"""Deprecated - use :mod:`tools.reference_data` instead."""
 
+from __future__ import annotations
+
+import warnings
 from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy import select
-import logging
-
 from db.models import Site
 
-logger = logging.getLogger(__name__)
+from .reference_data import ReferenceDataManager
 
+_manager = ReferenceDataManager()
 
 
 async def get_site(db: AsyncSession, site_id: int) -> Site | None:
-    """Fetch a site by ID.
-
-    Args:
-        db: Async SQLAlchemy session used for the query.
-        site_id: Identifier of the site to retrieve.
-
-    Returns:
-        The ``Site`` instance if found, otherwise ``None``.
-    """
-
-    return await db.get(Site, site_id)
+    warnings.warn(
+        "get_site is deprecated; use ReferenceDataManager.get_site",
+        DeprecationWarning,
+        stacklevel=2,
+    )
+    return await _manager.get_site(db, site_id)
 
 
 async def list_sites(db: AsyncSession, skip: int = 0, limit: int = 10) -> list[Site]:
-    """Return a paginated list of sites.
-
-    Args:
-        db: Async SQLAlchemy session used for the query.
-        skip: Number of records to skip from the beginning.
-        limit: Maximum number of records to return.
-
-    Returns:
-        A list of ``Site`` instances.
-    """
-
-    result = await db.execute(
-        select(Site).order_by(Site.ID).offset(skip).limit(limit)
+    warnings.warn(
+        "list_sites is deprecated; use ReferenceDataManager.list_sites",
+        DeprecationWarning,
+        stacklevel=2,
     )
-    return list(result.scalars().all())
+    return await _manager.list_sites(db, skip=skip, limit=limit)
