+++ tools/status_tools.py
@@ -1,14 +1,20 @@
-"""Helpers for listing ticket statuses."""
+"""Deprecated - use :mod:`tools.reference_data` instead."""
 
+from __future__ import annotations
+
+import warnings
 from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy import select
-import logging
-
 from db.models import TicketStatus
 
-logger = logging.getLogger(__name__)
+from .reference_data import ReferenceDataManager
+
+_manager = ReferenceDataManager()
 
 
 async def list_statuses(db: AsyncSession) -> list[TicketStatus]:
-    result = await db.execute(select(TicketStatus))
-    return list(result.scalars().all())
+    warnings.warn(
+        "list_statuses is deprecated; use ReferenceDataManager.list_statuses",
+        DeprecationWarning,
+        stacklevel=2,
+    )
+    return await _manager.list_statuses(db)
